Updated Upload of my files
So I wrote the files separately and from what I can tell on my desktop
they seem to be linked correctly (and were in the same folder on
desktop).

Basically just wanting to check it work and start allowing for linking
all the pages together.

I’ve tweaked the things I wanted to tweak:
- Colour background
- Updated social media buttons
- Embedded a youtube video (Fingers crossed it work)